Significant reformatting to improve the printed page
I was unsatisfied with how ABC.js embeds all stanza text within the SVG of the
last line of music. This breaks pagination. In particular, it causes the last
line of the music to be pulled onto the next page with the lyrics. Furthermore,
if the lyrics are longer than one page, they are clipped.

This change moves all lyrics out of the ABC and into the HTML.